# github-repo/openapi-schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EvokeEssence Exchange API",
    "description": "API schema for EvokeEssence cryptocurrency exchange",
    "version": "1.0.0",
    "contact": {
      "name": "EvokeEssence s.r.o",
      "url": "https://evo-exchange.com",
      "email": "contact@evo-exchange.com"
    }
  },
  "servers": [
    {
      "url": "https://evo-exchange.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/password-reset/request": {
      "post": {
        "summary": "Request password reset",
        "description": "Requests a password reset for a user account",
        "operationId": "requestPasswordReset",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email of the user requesting password reset",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset request processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "If your email is registered, you will receive a verification code shortly"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Email is required"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "An error occurred while processing your request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/password-reset/verify": {
      "post": {
        "summary": "Verify password reset",
        "description": "Verifies the password reset code and sets a new password",
        "operationId": "verifyPasswordReset",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "code",
                  "newPassword"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email of the user",
                    "example": "user@example.com"
                  },
                  "code": {
                    "type": "string",
                    "description": "Verification code received by email",
                    "example": "123456"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "New password",
                    "example": "SecurePassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Password has been reset successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired verification code"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "An error occurred while processing your request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/test": {
      "get": {
        "summary": "Test chat service",
        "description": "Tests if the chat service is properly configured",
        "operationId": "testChatService",
        "tags": [
          "Chat"
        ],
        "responses": {
          "200": {
            "description": "Chat service test successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "message": {
                      "type": "string",
                      "example": "The chat service is working correctly"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Chat service error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Chat service is not configured"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "Chat with AI assistant",
        "description": "Send a message to the AI assistant",
        "operationId": "chatWithAI",
        "tags": [
          "Chat"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content",
                  "sessionId"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "Message content",
                    "example": "What cryptocurrencies do you support?"
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Unique session identifier",
                    "example": "session-123456"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional metadata",
                    "example": {
                      "source": "web"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "message": {
                      "type": "string",
                      "example": "We support Bitcoin (BTC), Ethereum (ETH), and Tether (USDT) among others."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid message format"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to process chat message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Express session cookie for authentication"
      }
    },
...
  },
...
}